refactor: split API route monolith into domain modules
Extract the 1636-line [[route]].ts into 7 focused domain modules
under functions/lib/ (types, shared, auth, users, projects, documents,
ai, voice). The route file becomes a ~230-line thin router that
imports and dispatches to handlers. No behavioral changes.
